fix linting

+++ integration/signing_test.go
@@ -28,7 +28,7 @@ const (
 )
 
 var (
-	mgrMaxReconcileCount = 3
+	mgrMaxReconcileCount = uint(3)
 	validSubject         = pkix.Name{
 		CommonName: "CA", Organization: []string{"SE"}, Country: []string{"US"}, Locality: []string{"Canary"},
 		OrganizationalUnit: []string{"Clients"},
@@ -69,10 +69,10 @@ func TestTaskSigningAndVerification(t *testing.T) {
 			responderVerifier       *jwtsigning.Verifier
 			expJobStatus            orbital.JobStatus
 			expTaskStatus           orbital.TaskStatus
-			expInitiatorSignCalls   int
-			expInitiatorVerifyCalls int
-			expResponderSignCalls   int
-			expResponderVerifyCalls int
+			expInitiatorSignCalls   uint
+			expInitiatorVerifyCalls uint
+			expResponderSignCalls   uint
+			expResponderVerifyCalls uint
 		}{
 			{
 				name:                    "task request should sign and verify",
@@ -326,10 +326,10 @@ type testCase struct {
 	hasher                  jwtsigning.Hasher
 	expJobStatus            orbital.JobStatus
 	expTaskStatus           orbital.TaskStatus
-	expInitiatorSignCalls   int
-	expInitiatorVerifyCalls int
-	expResponderSignCalls   int
-	expResponderVerifyCalls int
+	expInitiatorSignCalls   uint
+	expInitiatorVerifyCalls uint
+	expResponderSignCalls   uint
+	expResponderVerifyCalls uint
 	expInitiatorVerifyErr   error
 	expResponderVerifyErr   error
 }
